+++ docs/changelog.md
@@ -23,6 +23,7 @@
   - Three positions for the name label: Left, Center and Right
   - You can set the verticall and horizontal offset of the name label
   - Added a new option to enable single portrait mode. In this mode, once the characters join the dialog, only one of them will be visible without the need of making them join and leave every time
+  - Added a simple fade in animation for dialogs. You can change how long it takes in the `Dialog Box` tab
   - New tab added: Audio
     - In the audio tab, you can select a sound or set of sounds to play while the text is being typed in the dialog. [[Tim Krief](https://github.com/timkrief)]